fix(fe): move component to upper level
