+++ apps/frontend/app/(client)/(main)/course/[courseId]/_components/SubmissionDetailModal.tsx
