+++ apps/frontend/app/(client)/(main)/course/[courseId]/assignment/[assignmentId]/@tabs/problem/_components/Columns.tsx
@@ -9,7 +9,7 @@ import { ErrorBoundary } from '@suspensive/react'
 import type { ColumnDef } from '@tanstack/react-table'
 import { Suspense, useState } from 'react'
 import { DetailButton } from '../../../../../_components/DetailButton'
-import { SubmissionDetailModal } from '../../../../../grade/_components/SubmissionDetailModal'
+import { SubmissionDetailModal } from '../../../../../_components/SubmissionDetailModal'
 
 export const columns = (
   assignment: AssignmentGrade
