+++ apps/frontend/app/admin/_components/table/DataTableMultiSelectFilter.tsx
@@ -55,7 +55,7 @@ export function DataTableMultiSelectFilter<TData, TValue>({
         <Button
           variant="outline"
           size={'sm'}
-          className="h-10 border hover:bg-gray-50"
+          className="h-10 rounded-full border hover:bg-gray-50"
         >
           <IoFilter className="mr-2 h-4 w-4" />
           <p className="font-bold">{title}</p>
